refactor(core): organize services package structure
