+++ src/main/java/pl/krywion/usosremastered/service/employee/EmployeeServiceImpl.java
@@ -4,10 +4,12 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import pl.krywion.usosremastered.dto.domain.EmployeeDto;
 import pl.krywion.usosremastered.dto.response.ServiceResponse;
+import pl.krywion.usosremastered.service.EmployeeService;
 import pl.krywion.usosremastered.service.employee.command.CreateEmployeeCommand;
 import pl.krywion.usosremastered.service.employee.command.DeleteEmployeeCommand;
 import pl.krywion.usosremastered.service.employee.command.EmployeeCommandHandler;
 import pl.krywion.usosremastered.service.employee.command.UpdateEmployeeCommand;
+import pl.krywion.usosremastered.service.employee.query.EmployeeQueryService;
 
 import java.util.List;
