+++ src/main/java/pl/krywion/usosremastered/service/student/StudentServiceImpl.java
@@ -4,10 +4,12 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import pl.krywion.usosremastered.dto.domain.StudentDto;
 import pl.krywion.usosremastered.dto.response.ServiceResponse;
+import pl.krywion.usosremastered.service.StudentService;
 import pl.krywion.usosremastered.service.student.command.CreateStudentCommand;
 import pl.krywion.usosremastered.service.student.command.DeleteStudentCommand;
 import pl.krywion.usosremastered.service.student.command.StudentCommandHandler;
 import pl.krywion.usosremastered.service.student.command.UpdateStudentCommand;
+import pl.krywion.usosremastered.service.student.query.StudentQueryService;
 
 
 import java.util.List;
