+++ src/main/java/pl/krywion/usosremastered/service/student/query/StudentQueryServiceImpl.java
@@ -1,4 +1,4 @@
-package pl.krywion.usosremastered.service.student;
+package pl.krywion.usosremastered.service.student.query;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
